Split multimap btree internals into tree_store and tighten visibility
Resolve the TODO on btree_base::BranchAccessor ("this should be
pub(super) and the multimap btree stuff should be moved into this
package") by factoring apart the multimap code so that the low-level
btree machinery lives under tree_store while the user-facing table API
stays at the crate's top level.

New module src/tree_store/multimap_btree.rs owns the on-page
serialization format (DynamicCollection, UntypedDynamicCollection,
DynamicCollectionType) and the tree-traversal helpers that touch raw
page layout: multimap_btree_stats, verify_tree_and_subtree_checksums,
relocate_subtrees, finalize_tree_and_subtree_checksums, and
UntypedMultiBtree. DynamicCollection::iter and iter_free_on_drop become
MultimapValue::from_collection and from_collection_free_on_drop, which
resolves the would-be circular dependency between the two modules.

With those helpers now inside tree_store, their only callers (btree.rs,
table_tree.rs, table_tree_base.rs, multimap_btree.rs) are inside the
module tree too. That lets a handful of btree_base types drop from
pub(crate) to pub(super): BranchAccessor, BranchMutator, LeafPageMut,
DEFERRED, and Checksum. The same audit found a broader set of
tree_store-internal items that can drop to pub(super): UntypedBtree,
UntypedBtreeMut, btree_stats, verify_tree_and_subtree_checksums,
relocate_subtrees, finalize_tree_and_subtree_checksums, UntypedMultiBtree,
and page_store::bitmap::BtreeBitmap. The corresponding re-exports are
removed from tree_store/mod.rs so the narrower visibility is actually
enforced at the module boundary.

BtreeStats and PagePath stay pub(crate): BtreeStats leaks out via the
return type of multimap_btree_stats (called from src/multimap_table.rs),
and PagePath appears in TableTreeMut::highest_index_pages, which
src/transactions.rs calls during compaction. LEAF, BRANCH, LeafAccessor,
and RawLeafBuilder also stay pub(crate) because MultimapTable::insert
and remove still build inline leaf blobs directly. BtreeHeader remains
pub(crate) for table.rs, db.rs, and transactions.rs.
