+++ src/tree_store/btree_mutator.rs
@@ -1,13 +1,13 @@
 use crate::tree_store::btree_base::{
     BRANCH, BranchAccessor, BranchBuilder, BranchMutator, Checksum, DEFERRED, LEAF, LeafAccessor,
-    LeafBuilder, LeafMutator,
+    LeafBuilder, LeafMutator, RawLeafBuilder,
 };
 use crate::tree_store::btree_mutator::DeletionResult::{
     DeletedBranch, DeletedLeaf, PartialBranch, PartialLeaf, Subtree,
 };
 use crate::tree_store::page_store::{Page, PageImpl, PageMut};
 use crate::tree_store::{
-    AccessGuardMutInPlace, BtreeHeader, PageHint, PageNumber, PageTrackerPolicy, RawLeafBuilder,
+    AccessGuardMutInPlace, BtreeHeader, PageHint, PageNumber, PageTrackerPolicy,
     TransactionalMemory,
 };
 use crate::types::{Key, Value};
