+++ src/tree_store/mod.rs
@@ -2,19 +2,17 @@ mod btree;
 mod btree_base;
 mod btree_iters;
 mod btree_mutator;
+mod multimap_btree;
 mod page_store;
 mod table_tree;
 mod table_tree_base;
 
-pub(crate) use btree::{
-    Btree, BtreeMut, BtreeStats, PagePath, RawBtree, UntypedBtree, UntypedBtreeMut, btree_stats,
-};
+pub(crate) use btree::{Btree, BtreeMut, BtreeStats, RawBtree};
+pub(crate) use btree_base::BtreeHeader;
 pub use btree_base::{AccessGuard, AccessGuardMut, AccessGuardMutInPlace};
-pub(crate) use btree_base::{
-    BRANCH, BranchAccessor, BranchMutator, BtreeHeader, Checksum, DEFERRED, LEAF, LeafAccessor,
-    LeafPageMut, RawLeafBuilder,
-};
+pub(crate) use btree_base::{BRANCH, LEAF, LeafAccessor, RawLeafBuilder};
 pub(crate) use btree_iters::{AllPageNumbersBtreeIter, BtreeExtractIf, BtreeRangeIter};
+pub(crate) use multimap_btree::{DynamicCollection, DynamicCollectionType, multimap_btree_stats};
 pub(crate) use page_store::ReadOnlyBackend;
 pub(crate) use page_store::{
     FILE_FORMAT_VERSION3, MAX_PAIR_LENGTH, MAX_VALUE_LENGTH, PAGE_SIZE, Page, PageHint, PageNumber,
